% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsa_manuscript.R
\name{apsa_manuscript}
\alias{apsa_manuscript}
\title{Political science manuscript format}
\usage{
apsa_manuscript(fig_width = 6.5, fig_height = 4.5, fig_crop = TRUE,
  highlight = "default", dev = "pdf", keep_tex = TRUE,
  latex_engine = "pdflatex", includes = NULL, md_extensions = NULL,
  pandoc_args = NULL)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for submitting a manuscript to a political science journal.
Adapted from
\href{http://www.apsanet.org/Portals/54/APSA\%20Files/publications/APSAStyleManual2006.pdf}{APSA Style Manual for Political Science},
and the submission instructions for \emph{APSA}, \emph{AJPS}, \emph{JOP}, and \emph{Political Analysis}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyManuscript.Rmd", template = "apsa_manuscript", package = "apsamanu")
}

}

