% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_manuscript.R
\name{ps_manuscript_pdf}
\alias{ps_manuscript_pdf}
\alias{ps_manuscript_word}
\title{Political Science PDF manuscript format}
\usage{
ps_manuscript_pdf(..., number_sections = FALSE, fig_caption = TRUE,
  dev = "pdf", highlight = NULL, latex_engine = "xelatex",
  citation_package = "natbib", pandoc_args = NULL, margin = 1,
  fontsize = 12, linestretch = 2, colorlinks = FALSE)

ps_manuscript_word(..., csl = NULL)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{csl}{Path to CSL file to use}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for submitting a manuscript to a political science journal.
Adapted from
\href{http://www.apsanet.org/Portals/54/APSA\%20Files/publications/APSAStyleManual2006.pdf}{APSA Style Manual for Political Science},
and the submission instructions for \emph{APSA}, \emph{AJPS}, \emph{JOP}, and \emph{Political Analysis}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyManuscript.Rmd", template = "ps_manuscript_pdf", package = "rmdmanuscript")
}
}

