% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_manuscript.R
\name{ps_manuscript_pdf}
\alias{ps_manuscript_pdf}
\alias{ps_manuscript_word}
\title{Political Science PDF manuscript format}
\usage{
ps_manuscript_pdf(..., keep_tex = TRUE)

ps_manuscript_word(..., csl = NULL)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{csl}{Path to CSL file to use}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for submitting a manuscript to a political science journal.
Adapted from
\href{http://www.apsanet.org/Portals/54/APSA\%20Files/publications/APSAStyleManual2006.pdf}{APSA Style Manual for Political Science},
and the submission instructions for \emph{APSA}, \emph{AJPS}, \emph{JOP}, and \emph{Political Analysis}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyManuscript.Rmd", template = "ps_manuscript_pdf", package = "rmdmanuscript")
}

}

